Only Focus one Table
Until now the targets and results table were focused, which caused a
weird behavior when a user scrolled through the tables. This is now
fixed by only focusing one table. The logic for which table is focused
depends on the `ActiveTargetIndex` variable, which means when the
variable is `-1` the targets table is focused and otherwise the results
table.