Add Notifications
It is now possible to get notified for failed checks or for checks which
took longer then expected. This feature can be enabled via the
`-notification` flag and the `-notification-threshold` flag. While the
first on is needed to enable notifications the seconde one can be used
to also send notifications when a check took longer then the configured
duration. These two new options can also be set in the configuration
file, for each target.