Remove GoReleaser and Use Custom GitHub Actions for Releases (#28)
Remove the GoReleaser GitHub Action to publish releases and replace it
with the Release Drafter GitHub Action to generate the changelog for a
release and a custom GitHub Action to build and publish the binaries.

This is done, because I use a similar setup in other repositories and
want to keep things consistent across my projects.